Abstract out entity sub feature reducers into own files
